requirements.txt
streamlit
websocket-client
requests
python-dotenv
